Add files via upload
with the scrape.js file: Open your web browser. Go to the human benchmark typing test and open the web browser inspector. go to console and paste the code from the file. Hit enter, then copy the return from the function.

With the parse.py file: paste the data from the scraper into the variable called `my_string` run it, and then copy the return without the comma at the end.

With the cheat.c: ensure you have the necessary libraries. then paste the return without the comma at the end into the list called key_list between the curly brackets. compile the file run it. you will have 10 seconds to switch open to the relevant tab it will type it out for you.

errors:
1. you set it up wrong
2. you don't have the necessary libraries.
3. the parser doesn't contain all of the non-alphabet characters (you have to add them if they aren't there)